remove_preprocessing_2
